+++ src/utils/logger.mock.ts
@@ -1,6 +1,8 @@
 import { Logger } from "./logger";
 
-export const getLoggerMock = (): jest.Mocked<Logger> => ({
+export type LoggerMock = jest.Mocked<Logger>;
+
+export const getLoggerMock = (): LoggerMock => ({
   error: jest.fn(),
   warn: jest.fn(),
   info: jest.fn(),